modified docker config and merged some files for code readibility
