+++ Gatoride Project/docker-compose.yml
@@ -7,7 +7,4 @@ services:
     build: ./flask
     ports:
       - 3001:3000
-  web:
-    build: ./react-app
-    ports:
-      - 3000:3000
+
