+++ Gatoride Project/flask/Dockerfile
@@ -14,5 +14,8 @@ COPY . /app
 # Expose the port used by flask
 EXPOSE 3000
 
+# Set the FLASK_APP environment variable to your Flask application file
+ENV FLASK_APP=run.py
+
 # Run the flask program
-CMD ["flask", "--app", "__init__", "run", "--debug", "--port=3000", "--host=0.0.0.0"]
+CMD ["flask", "run", "--debug", "--port=3000", "--host=0.0.0.0"]
