+++ Gatoride Project/flask/app/__init__.py
@@ -1,10 +1,10 @@
-from flask import Flask , jsonify
+from flask import Flask, jsonify
 from .models import db
 from .routes import configure_routes
 
 def create_app():
     app = Flask(__name__)
-    app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///C:/Users/joeli/code/Gatoride/persons.db'
+    app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///persons.db'
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
     db.init_app(app)
     with app.app_context():
@@ -18,4 +18,3 @@ def return_status(status):
         return "successful"
     else:
         return "unsuccessful"
-
